Adding ressy support
Flixel can be compiled with or without ressy support or the flixel assets. Just add -D ressy to add the functions specific to ressy, and -D flixelAssets to activate ressy and some images or font originally used by Flixel.

Unfortunately, the loading of the assets doesn't work yet when using Loddy, but works fine with Eddy.